Fixed observer

+++ Task1/src/by/andruhovich/task/observer/Observer.java
@@ -3,5 +3,5 @@ package by.andruhovich.task.observer;
 import by.andruhovich.task.entity.Quadrangle;
 
 public interface Observer {
-    public void update(Integer id, Quadrangle quadrangle);
+    public void update(long id, Quadrangle quadrangle);
 }
